netboot/dhcp/requirements.txt
absl-py==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
more-itertools==6.0.0
pluggy==0.9.0
py==1.8.0
pytest==4.3.1
six==1.12.0
